Restructured firebase in the picturepoint src file. Added efficient getUser and getPhoto functions(not implemented in react) #66
