+++ picturepoint/src/Firebase/functions/auth.js
