+++ picturepoint/src/Firebase/index.js
@@ -1,0 +1,4 @@
+import * as auth from "./functions/auth";
+import * as firebase from "./functions/firebase";
+
+export { auth, firebase };
