+++ picturepoint/src/Layouts/AuthHeader.jsx
@@ -4,7 +4,7 @@ import { Link } from "react-router-dom";
 import * as routes from "../Routes/routes";
 
 //Firebase
-import { auth } from "../Login/firebase";
+import { auth } from "../Firebase/index";
 
 //Material UI
 import { makeStyles } from '@material-ui/core/styles';
