+++ picturepoint/src/Login/Login.jsx
@@ -1,6 +1,6 @@
 //React
 import React from 'react';
-import { BrowserRouter as Router, Route } from "react-router-dom";
+import { Route } from "react-router-dom";
 
 //Routes
 import * as routes from "../Routes/routes";
