+++ picturepoint/src/Login/components/PasswordChange.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 
-import { auth } from "../firebase";
+import { auth } from "../../Firebase/index";
 
 const byPropKey = (propertyName, value) => () => ({
   [propertyName]: value
