+++ picturepoint/src/Login/components/withAuthentication.js
@@ -1,7 +1,7 @@
 import React from "react";
 
 import AuthUserContext from "./AuthUserContext";
-import { firebase } from "../firebase";
+import { firebase } from "../../Firebase/index";
 
 const withAuthentication = Component =>
   class WithAuthentication extends React.Component {
