--- picturepoint/src/Login/firebase/index.js
@@ -1,4 +1,0 @@
-import * as auth from "./auth";
-import * as firebase from "./firebase";
-
-export { auth, firebase };
